+++ CHANGELOG.md
@@ -40,7 +40,7 @@ Emojis for the following are chosen based on [gitmoji](https://gitmoji.dev/).
 - Fixes an issue with iPads where semicolon keys also had apostrophes.
 - Fixes an issue where number keys weren't able to trigger the double space period shortcut.
 - Fixes an issue where hold-to-select keys wouldn't return to their original color.
-<!-- - Fixes an issue where more than one singular gender wasn't being assigned to German nouns in the formatting process.  -->
+- Fixes an issue where more than one singular gender wasn't being assigned to German nouns in the formatting process.
 
 ### ♻️ Code Refactoring