+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Emojis for the following are chosen based on [gitmoji](https://gitmoji.dev/).
 ### 🌐 New Keyboards
 
 - Adds Russian, French, Portuguese and Swedish keyboards.
-- Verb support is currently limited, but noun annotation, plural and `beta` translation are available.
+- Verb conjugation support is currently limited, but noun annotation, plural and `beta` translation are available.
 
 ### ✨ New Features